+++ controls-ports-ktor/src/jvmMain/kotlin/space/kscience/controls/ports/KtorUdpPort.kt
@@ -4,9 +4,9 @@ import io.ktor.network.selector.ActorSelectorManager
 import io.ktor.network.sockets.*
 import io.ktor.utils.io.ByteWriteChannel
 import io.ktor.utils.io.consumeEachBufferRange
-import io.ktor.utils.io.core.Closeable
 import io.ktor.utils.io.writeAvailable
 import kotlinx.coroutines.*
+import space.kscience.controls.api.LifecycleState
 import space.kscience.dataforge.context.Context
 import space.kscience.dataforge.context.Factory
 import space.kscience.dataforge.meta.Meta
@@ -24,7 +24,7 @@ public class KtorUdpPort internal constructor(
     public val localHost: String = "localhost",
     coroutineContext: CoroutineContext = context.coroutineContext,
     socketOptions: SocketOptions.UDPSocketOptions.() -> Unit = {},
-) : AbstractAsynchronousPort(context, meta, coroutineContext), Closeable {
+) : AbstractAsynchronousPort(context, meta, coroutineContext) {
 
     override fun toString(): String = "port[udp:$remoteHost:$remotePort]"
 
@@ -58,13 +58,13 @@ public class KtorUdpPort internal constructor(
         writeChannel.await().writeAvailable(data)
     }
 
-    override val isOpen: Boolean
-        get() = listenerJob?.isActive == true
+    override val lifecycleState: LifecycleState
+        get() = if(listenerJob?.isActive == true) LifecycleState.STARTED else LifecycleState.STOPPED
 
-    override fun close() {
+    override suspend fun stop() {
         listenerJob?.cancel()
         futureSocket.cancel()
-        super.close()
+        super.stop()
     }
 
     public companion object : Factory<AsynchronousPort> {
@@ -101,7 +101,7 @@ public class KtorUdpPort internal constructor(
         /**
          * Create and open UDP port
          */
-        public fun open(
+        public suspend fun start(
             context: Context,
             remoteHost: String,
             remotePort: Int,
@@ -117,7 +117,7 @@ public class KtorUdpPort internal constructor(
             localHost,
             coroutineContext,
             socketOptions
-        ).apply { open() }
+        ).apply { start() }
 
         override fun build(context: Context, meta: Meta): AsynchronousPort {
             val remoteHost by meta.string { error("Remote host is not specified") }